refactor: style fixes
